url params fixed for single profile
