+++ content-feed.html
@@ -103,7 +103,7 @@
               <svg class="bi pe-none me-2" width="30" height="24"><use xlink:href="#bootstrap"></use></svg>
               <span class="fs-5 fw-semibold">Friends online</span>
             </a>
-            
+
             <div class="list-group list-group-flush border-bottom scrollarea" id="friendsOnline">
 
             </div>
@@ -111,7 +111,7 @@
         </div>
     </main>
     <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-    <script type="module" src="/js/components/requests/profile.mjs"></script>
+    <script type="module" src="/js/components/requests/getProfiles.mjs"></script>
     <script tpye="module" src="/js/components/requests/createPost.mjs"></script>
     <script type="module" src="/js/components/requests/viewFeed.mjs"></script>
 </body>
