+++ profile.html
@@ -95,7 +95,7 @@
         </div>
     </main>
     <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="/js/components/requests/profile.mjs" type="module"></script>
-    <script type="module" src="/js/components/requests/viewFeed.mjs"></script>
+    <script src="/js/components/requests/getProfiles.mjs" type="module"></script>
+    <script type="module" src="/js/components/requests/singleProfile.mjs"></script>
 </body>
 </html>
